Remove .config/lvim from runtime path
I think this is a bad idea because it can conflict with our code base and loadfile() is already a good replacement.
If we want to do this, I would suggest we move the source code under a lvim/ dir.
The parent dir would act as a namespace and allow us to write require('lvim.module_name').